+++ test/Fixtures/Mariadb/SimpleTestCases/NetworkTypes/expectedSchema.php
@@ -2,12 +2,9 @@
 
 declare(strict_types=1);
 
-use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Database\Connection\OneToManyConnection;
 use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Database\Table;
 use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Database\Table\Column;
 use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Database\Table\Index\PrimaryKey;
-use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Database\Table\Index\SimpleIndex;
-use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Database\Table\Index\UniqueIndex;
 use Pongee\DatabaseSchemaVisualization\DataObject\Sql\Schema;
 
 return new Schema()
@@ -18,15 +15,6 @@ return new Schema()
                     'id',
                     'UUID',
                     [],
-                    'NOT NULL',
-                    ''
-                )
-            )
-            ->addColumn(
-                new Column(
-                    'hostname',
-                    'varchar',
-                    [255],
                     'NOT NULL',
                     ''
                 )
@@ -50,55 +38,4 @@ return new Schema()
                 )
             )
             ->setPrimaryKey(new PrimaryKey(['id']))
-            ->addUniqueIndex(
-                new UniqueIndex(
-                    'uq_hostname',
-                    ['hostname']
-                )
-            )
-    )
-    ->addTable(
-        new Table('device_interface')
-            ->addColumn(
-                new Column(
-                    'id',
-                    'UUID',
-                    [],
-                    'NOT NULL',
-                    ''
-                )
-            )
-            ->addColumn(
-                new Column(
-                    'device_id',
-                    'UUID',
-                    [],
-                    'NOT NULL',
-                    ''
-                )
-            )
-            ->addColumn(
-                new Column(
-                    'address',
-                    'INET6',
-                    [],
-                    'NOT NULL',
-                    ''
-                )
-            )
-            ->setPrimaryKey(new PrimaryKey(['id']))
-            ->addSimpleIndex(
-                new SimpleIndex(
-                    'fk_device_id',
-                    ['device_id']
-                )
-            )
-    )
-    ->addConnection(
-        new OneToManyConnection(
-            'device_interface',
-            'device',
-            ['device_id'],
-            ['id']
-        )
     );