+++ test/bootstrap.php
@@ -2,4 +2,4 @@
 
 declare(strict_types=1);
 
-define('FIXTURES_DIRECTORY', __DIR__ . '/Fixtures/');
+const FIXTURES_DIRECTORY = __DIR__ . '/Fixtures/';